+++ packages/core/index.js
@@ -1,13 +1,5 @@
-const htmlparser2patch = require('./src/patches/htmlparser2');
-const markdownItEscapeSpecialTags = require('./src/lib/markdown-it/markdown-it-escape-special-tags');
-const Parser = require('./src/Parser');
-
-function ignoreTags(tagsToIgnore) {
-  htmlparser2patch.injectIgnoreTags(tagsToIgnore);
-  markdownItEscapeSpecialTags.injectTags(tagsToIgnore);
-}
+const Site = require('./src/Site');
 
 module.exports = {
-  Parser,
-  ignoreTags,
+  Site,
 };
