+++ packages/core/src/Site/index.js
@@ -1,4 +1,4 @@
-const cheerio = require('cheerio'); require('@markbind/core/src/patches/htmlparser2');
+const cheerio = require('cheerio'); require('../patches/htmlparser2');
 const fs = require('fs-extra-promise');
 const ghpages = require('gh-pages');
 const ignore = require('ignore');
@@ -7,19 +7,23 @@ const Promise = require('bluebird');
 const ProgressBar = require('progress');
 const walkSync = require('walk-sync');
 
-const markbind = require('@markbind/core');
-const Page = require('@markbind/core/src/Page');
-const VariablePreprocessor = require('@markbind/core/src/preprocessors/VariablePreprocessor');
-const Template = require('@markbind/core/template/template');
+const SiteConfig = require('./SiteConfig');
+const Page = require('../Page');
+const { ignoreTags } = require('../patches');
+const VariablePreprocessor = require('../preprocessors/VariablePreprocessor');
+const Template = require('../../template/template');
 
-const FsUtil = require('@markbind/core/src/utils/fsUtil');
-const njUtil = require('@markbind/core/src/utils/nunjuckUtils');
-const utils = require('@markbind/core/src/utils');
+const FsUtil = require('../utils/fsUtil');
+const njUtil = require('../utils/nunjuckUtils');
+const delay = require('../utils/delay');
+const logger = require('../utils/logger');
+const utils = require('../utils');
+
 const {
   LAYOUT_DEFAULT_NAME,
   LAYOUT_FOLDER_PATH,
   PLUGIN_SITE_ASSET_FOLDER_NAME,
-} = require('@markbind/core/src/constants');
+} = require('../constants');
 
 const _ = {};
 _.difference = require('lodash/difference');
@@ -38,11 +42,7 @@ _.uniq = require('lodash/uniq');
 const url = {};
 url.join = path.posix.join;
 
-const delay = require('./util/delay');
-const logger = require('./util/logger');
-const SiteConfig = require('./SiteConfig');
-
-const CLI_VERSION = require('../package.json').version;
+const MARKBIND_VERSION = require('../../package.json').version;
 
 const {
   ABOUT_MARKDOWN_FILE,
@@ -111,7 +111,7 @@ const TOP_NAV_DEFAULT = '<header><navbar placement="top" type="inverse">\n'
   + '  </li>\n'
   + '</navbar></header>';
 
-const MARKBIND_LINK_HTML = `<a href='${MARKBIND_WEBSITE_URL}'>MarkBind ${CLI_VERSION}</a>`;
+const MARKBIND_LINK_HTML = `<a href='${MARKBIND_WEBSITE_URL}'>MarkBind ${MARKBIND_VERSION}</a>`;
 
 class Site {
   constructor(rootPath, outputPath, onePagePath, forceReload = false, siteConfigPath = SITE_CONFIG_NAME) {
@@ -120,11 +120,11 @@ class Site {
     this.tempPath = path.join(rootPath, TEMP_FOLDER_NAME);
 
     // MarkBind assets to be copied
-    this.siteAssetsSrcPath = path.resolve(__dirname, '../packages/core', SITE_ASSET_FOLDER_NAME);
+    this.siteAssetsSrcPath = path.resolve(__dirname, '../..', SITE_ASSET_FOLDER_NAME);
     this.siteAssetsDestPath = path.join(outputPath, TEMPLATE_SITE_ASSET_FOLDER_NAME);
 
     // Page template path
-    this.pageTemplatePath = path.join(__dirname, PAGE_TEMPLATE_NAME);
+    this.pageTemplatePath = path.join(__dirname, '../Page', PAGE_TEMPLATE_NAME);
     this.pageTemplate = njUtil.compile(fs.readFileSync(this.pageTemplatePath, 'utf8'));
     this.pages = [];
 
@@ -975,7 +975,7 @@ class Site {
       });
     });
 
-    markbind.ignoreTags(tagsToIgnore);
+    ignoreTags(tagsToIgnore);
 
     Page.htmlBeautifyOptions = {
       indent_size: 2,
@@ -1139,7 +1139,7 @@ class Site {
    * Copies Font Awesome assets to the assets folder
    */
   copyFontAwesomeAsset() {
-    const faRootSrcPath = path.join(__dirname, '..', 'node_modules', '@fortawesome', 'fontawesome-free');
+    const faRootSrcPath = path.dirname(require.resolve('@fortawesome/fontawesome-free/package.json'));
     const faCssSrcPath = path.join(faRootSrcPath, 'css', 'all.min.css');
     const faCssDestPath = path.join(this.siteAssetsDestPath, 'fontawesome', 'css', 'all.min.css');
     const faFontsSrcPath = path.join(faRootSrcPath, 'webfonts');
@@ -1152,8 +1152,7 @@ class Site {
    * Copies Octicon assets to the assets folder
    */
   copyOcticonsAsset() {
-    const octiconsRootSrcPath = path.join(__dirname, '..', 'node_modules', '@primer', 'octicons', 'build');
-    const octiconsCssSrcPath = path.join(octiconsRootSrcPath, 'build.css');
+    const octiconsCssSrcPath = require.resolve('@primer/octicons/build/build.css');
     const octiconsCssDestPath = path.join(this.siteAssetsDestPath, 'css', 'octicons.css');
 
     return fs.copyAsync(octiconsCssSrcPath, octiconsCssDestPath);
@@ -1163,8 +1162,7 @@ class Site {
    * Copies components.min.js bundle to the assets folder
    */
   copyComponentsAsset() {
-    const componentsSrcPath = path.join(__dirname, '..', 'packages', 'vue-components', 'dist',
-                                        'components.min.js');
+    const componentsSrcPath = require.resolve('@markbind/vue-components/dist/components.min.js');
     const componentsDestPath = path.join(this.siteAssetsDestPath, 'js', 'components.min.js');
 
     return fs.copyAsync(componentsSrcPath, componentsDestPath);
