+++ packages/core/test/unit/Site.test.js
@@ -6,6 +6,7 @@ const Site = require('../../src/Site');
 const {
   ABOUT_MD_DEFAULT,
   ASSETS,
+  EXTERNAL_ASSETS,
   FOOTER_MD_DEFAULT,
   HEADER_MD_DEFAULT,
   INDEX_MD_DEFAULT,
@@ -26,23 +27,19 @@ jest.mock('fs');
 jest.mock('walk-sync');
 jest.mock('@markbind/core/src/Page');
 jest.mock('gh-pages');
-jest.mock('../../src/util/logger');
 
 afterEach(() => fs.vol.reset());
 
 test('Site Generate builds the correct amount of assets', async () => {
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     'inner/site.json': SITE_JSON_DEFAULT,
 
     ...ASSETS,
 
     'packages/vue-components/dist/components.min.js': '',
 
-    'node_modules/@fortawesome/fontawesome-free/css/all.min.css': '',
-    'node_modules/@fortawesome/fontawesome-free/webfonts/font1.svg': '',
-    'node_modules/@fortawesome/fontawesome-free/webfonts/font2.ttf': '',
-    'node_modules/@primer/octicons/build/build.css': '',
+    ...EXTERNAL_ASSETS,
 
     'inner/_markbind/layouts/default/scripts.js': '',
     'inner/_markbind/layouts/default/styles.css': '',
@@ -92,7 +89,7 @@ test('Site Generate builds the correct amount of assets', async () => {
 test('Site Init with invalid template fails', async () => {
   // Mock default template in MemFS without site config
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     [getDefaultTemplateFileFullPath('index.md')]: INDEX_MD_DEFAULT,
   };
 
@@ -111,7 +108,7 @@ test('Site Init does not overwrite existing files', async () => {
   // Mock default template in MemFS
   const json = {
     'index.md': EXISTING_INDEX_MD,
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     [getDefaultTemplateFileFullPath('index.md')]: INDEX_MD_DEFAULT,
     [getDefaultTemplateFileFullPath('site.json')]: SITE_JSON_DEFAULT,
   };
@@ -125,7 +122,7 @@ test('Site Init does not overwrite existing files', async () => {
 test('Site Init in existing directory generates correct assets', async () => {
   // Mock default template in MemFS
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     ...DEFAULT_TEMPLATE_FILES,
   };
 
@@ -175,7 +172,7 @@ test('Site Init in existing directory generates correct assets', async () => {
 test('Site Init in directory which does not exist generates correct assets', async () => {
   // Mock default template in MemFS
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     ...DEFAULT_TEMPLATE_FILES,
   };
   fs.vol.fromJSON(json, '');
@@ -227,7 +224,7 @@ test('Site Init in directory which does not exist generates correct assets', asy
 
 test('Site baseurls are correct for sub nested subsites', async () => {
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     'site.json': SITE_JSON_DEFAULT,
     'sub/site.json': SITE_JSON_DEFAULT,
     'sub/sub/site.json': SITE_JSON_DEFAULT,
@@ -244,7 +241,7 @@ test('Site baseurls are correct for sub nested subsites', async () => {
 
 test('Site removeAsync removes the correct asset', async () => {
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     '_site/toRemove.jpg': '',
     '_site/dontRemove.png': '',
     'toRemove.html': '',
@@ -259,7 +256,7 @@ test('Site removeAsync removes the correct asset', async () => {
 
 test('Site read site config for default', async () => {
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     'site.json': SITE_JSON_DEFAULT,
   };
   fs.vol.fromJSON(json, '');
@@ -297,7 +294,7 @@ test('Site read site config for custom site config', async () => {
     enableSearch: true,
   };
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     'site.json': JSON.stringify(customSiteJson),
   };
   fs.vol.fromJSON(json, '');
@@ -314,7 +311,7 @@ test('Site read site config for custom site config', async () => {
 
 test('Site resolves variables referencing other variables', async () => {
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     'site.json': SITE_JSON_DEFAULT,
     '_markbind/variables.md':
     '<variable name="level1">variable</variable>'
@@ -341,7 +338,7 @@ test('Site resolves variables referencing other variables', async () => {
 
 test('Site read correct user defined variables', async () => {
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     'site.json': SITE_JSON_DEFAULT,
     'sub/site.json': SITE_JSON_DEFAULT,
     'sub/sub/site.json': SITE_JSON_DEFAULT,
@@ -384,7 +381,7 @@ test('Site read correct user defined variables', async () => {
 test('Site convert generates correct assets', async () => {
   // Mock default template in MemFS
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     ...DEFAULT_TEMPLATE_FILES,
   };
   fs.vol.fromJSON(json, '');
@@ -423,7 +420,7 @@ test('Site convert generates correct assets', async () => {
 test('Site convert with custom _Footer.md, no _Sidebar.md, README.md generates correct assets', async () => {
   // Mock default template in MemFS
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     ...DEFAULT_TEMPLATE_FILES,
     // Custom Footer and README
     'inner/_Footer.md': 'Custom footer.',
@@ -465,7 +462,7 @@ test('Site convert with custom _Footer.md, no _Sidebar.md, README.md generates c
 
 test('Site deploys with default settings', async () => {
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     'site.json': SITE_JSON_DEFAULT,
     _site: {},
   };
@@ -485,7 +482,7 @@ test('Site deploys with custom settings', async () => {
     branch: 'master',
   };
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     'site.json': JSON.stringify(customConfig),
     _site: {},
   };
@@ -499,7 +496,7 @@ test('Site deploys with custom settings', async () => {
 
 test('Site should not deploy without a built site', async () => {
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     'site.json': SITE_JSON_DEFAULT,
   };
   fs.vol.fromJSON(json, '');
@@ -533,7 +530,7 @@ describe('Site deploy with Travis', () => {
     process.env.TRAVIS_REPO_SLUG = 'TRAVIS_USER/TRAVIS_REPO';
 
     const json = {
-      'src/page.njk': PAGE_NJK,
+      ...PAGE_NJK,
       'site.json': SITE_JSON_DEFAULT,
       _site: {},
     };
@@ -553,7 +550,7 @@ describe('Site deploy with Travis', () => {
     const customRepoConfig = JSON.parse(SITE_JSON_DEFAULT);
     customRepoConfig.deploy.repo = 'https://github.com/USER/REPO.git';
     const json = {
-      'src/page.njk': PAGE_NJK,
+      ...PAGE_NJK,
       'site.json': JSON.stringify(customRepoConfig),
       _site: {},
     };
@@ -570,7 +567,7 @@ describe('Site deploy with Travis', () => {
     process.env.TRAVIS_REPO_SLUG = 'TRAVIS_USER/TRAVIS_REPO.github.io';
 
     const json = {
-      'src/page.njk': PAGE_NJK,
+      ...PAGE_NJK,
       'site.json': SITE_JSON_DEFAULT,
       _site: {},
     };
@@ -583,7 +580,7 @@ describe('Site deploy with Travis', () => {
 
   test('Site deploy -t/--travis should not deploy if not in Travis', async () => {
     const json = {
-      'src/page.njk': PAGE_NJK,
+      ...PAGE_NJK,
       'site.json': SITE_JSON_DEFAULT,
       _site: {},
     };
@@ -598,7 +595,7 @@ describe('Site deploy with Travis', () => {
     process.env.TRAVIS = true;
 
     const json = {
-      'src/page.njk': PAGE_NJK,
+      ...PAGE_NJK,
       'site.json': SITE_JSON_DEFAULT,
       _site: {},
     };
@@ -616,7 +613,7 @@ describe('Site deploy with Travis', () => {
     const invalidRepoConfig = JSON.parse(SITE_JSON_DEFAULT);
     invalidRepoConfig.deploy.repo = 'INVALID_GITHUB_REPO';
     const json = {
-      'src/page.njk': PAGE_NJK,
+      ...PAGE_NJK,
       'site.json': JSON.stringify(invalidRepoConfig),
       _site: {},
     };
@@ -734,7 +731,7 @@ siteJsonResolvePropertiesTestCases.forEach((testCase) => {
       },
     };
     const json = {
-      'src/page.njk': PAGE_NJK,
+      ...PAGE_NJK,
       'index.md': '',
     };
     fs.vol.fromJSON(json, '');
@@ -770,7 +767,7 @@ test('Site config throws error on duplicate page src', async () => {
     },
   };
   const json = {
-    'src/page.njk': PAGE_NJK,
+    ...PAGE_NJK,
     'index.md': '',
   };
   fs.vol.fromJSON(json, '');
